fix: MongoDB connection string contained invalid tokens
The `authConnectionString` was allowed to contain `{database}` as a variable which would be replaced with the correct database. However, Jekyll (which is powering the chart-releaser w/ Github pages) does not work correctly with curly braces within these YAML templates.

This is fixed by replacing this variable within the oAuth service itself.